+++ lib/PuppeteerSharp/EmulationManager.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
+using PuppeteerSharp.Media;
 
 namespace PuppeteerSharp
 {
